+++ app/controllers/projects_controller.rb
@@ -69,9 +69,9 @@ class ProjectsController < ApplicationController
   def project_params
     params[:project] ||= { blank: "1" }
     parse_date_if_key_present(:project, :launch_date)
-    params.require(:project).permit(
+    params.fetch(:project, {}).permit(
       :name, :slug, :access_token, :short_description, :consent, :theme,
-      :launch_date, :published
+      :launch_date, :published, :slice_site_id, :code_prefix
     )
   end
 end