+++ config/application.rb
@@ -40,5 +40,8 @@ module MyApnea
     config.autoload_paths << Rails.root.join("app", "models", "admin")
     config.autoload_paths << Rails.root.join("app", "models", "reports")
     config.autoload_paths << Rails.root.join("app", "models", "surveys")
+
+    # Add Slice models to autoload_paths.
+    config.autoload_paths << Rails.root.join("app", "models", "slice")
   end
 end