+++ db/migrate/20171204221119_add_slice_site_id_to_projects.rb
@@ -1,0 +1,5 @@
+class AddSliceSiteIdToProjects < ActiveRecord::Migration[5.2]
+  def change
+    add_column :projects, :slice_site_id, :integer
+  end
+end