+++ db/migrate/20171204221739_add_code_prefix_to_projects.rb
@@ -1,0 +1,5 @@
+class AddCodePrefixToProjects < ActiveRecord::Migration[5.2]
+  def change
+    add_column :projects, :code_prefix, :string
+  end
+end